NewPassword refactor 1 + navigation
